move cache into utils
